refactor(structs): add data app to hold core and worker structs
