+++ apps/data/mix.exs
@@ -12,25 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-defmodule FunlessWorker.MixProject do
+defmodule Data.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :worker,
-      version: "0.4.0",
+      app: :data,
+      version: "0.1.0",
       build_path: "../../_build",
       config_path: "../../config/config.exs",
       deps_path: "../../deps",
       lockfile: "../../mix.lock",
       elixir: "~> 1.14",
-      elixirc_paths: elixirc_paths(Mix.env()),
       start_permanent: Mix.env() == :prod,
-      rustler_crates: [
-        fn_wasm: [
-          mode: if(Mix.env() == :dev, do: :debug, else: :release)
-        ]
-      ],
       deps: deps(),
       aliases: aliases()
     ]
@@ -39,34 +33,19 @@ defmodule FunlessWorker.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      extra_applications: [:logger, :inets, :os_mon],
-      mod: {Worker.Application, []}
+      extra_applications: [:logger]
     ]
   end
 
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
-      {:data, in_umbrella: true},
-      {:rustler, "~> 0.26.0"},
-      {:jason, "~> 1.3"},
-      {:libcluster, "~> 3.3"},
-      {:logger_file_backend, "~> 0.0.13"},
-      {:telemetry, "~> 1.1.0"},
-      {:telemetry_poller, "~> 1.0"},
-      {:prom_ex, git: "https://github.com/akoutmos/prom_ex.git"},
-
-      # Dev deps
-      {:mox, "~> 1.0", only: :test}
+      # {:dep_from_hexpm, "~> 0.3.0"},
+      # {:dep_from_git, git: "https://github.com/elixir-lang/my_dep.git", tag: "0.1.0"},
+      # {:sibling_app_in_umbrella, in_umbrella: true}
     ]
   end
 
-  defp elixirc_paths(:test), do: ["test/support", "lib"]
-  defp elixirc_paths(_), do: ["lib"]
-
-  # Aliases are shortcuts or tasks specific to the current project.
-  #
-  # See the documentation for `Mix` for more info on aliases.
   defp aliases do
     [
       setup: ["deps.get"]
