+++ apps/worker/lib/worker/domain/cleanup_resource.ex
@@ -16,7 +16,7 @@ defmodule Worker.Domain.CleanupResource do
   @moduledoc """
   Contains functions used to remove function runtimes. Side effects (e.g. docker interaction) are delegated to ports and adapters.
   """
-  alias Worker.Domain.FunctionStruct
+  alias Data.FunctionStruct
   alias Worker.Domain.Ports.ResourceCache
   alias Worker.Domain.Ports.Runtime.Cleaner
